# internal/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "CRUD for servers, requests, and deliveries with rate-limited worker.",
        "title": "Sukumad API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "roles"
                ],
                "summary": "List roles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Role"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Create role",
                "parameters": [
                    {
                        "description": "Role payload",
                        "name": "role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Role"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.Role"
                        }
                    }
                }
            }
        },
        "/api/roles/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Update role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role payload",
                        "name": "role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Role"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "roles"
                ],
                "summary": "Delete role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/roles/{id}/permissions": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Grant permission to role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Permission grant",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GrantRolePermission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Revoke permission from role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Permission revoke",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.GrantRolePermission"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/servers": {
            "get": {
                "description": "List servers with pagination",
                "tags": [
                    "servers"
                ],
                "summary": "List servers",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIError"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "servers"
                ],
                "summary": "Create server",
                "parameters": [
                    {
                        "description": "Server payload",
                        "name": "server",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Server"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.Server"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIError"
                        }
                    }
                }
            }
        },
        "/api/servers/{id}": {
            "get": {
                "tags": [
                    "servers"
                ],
                "summary": "Get server",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Server"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.APIError"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "servers"
                ],
                "summary": "Update server",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Server payload",
                        "name": "server",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Server"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Server"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIError"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "servers"
                ],
                "summary": "Delete server",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "types.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "bad request"
                }
            }
        },
        "types.GrantRolePermission": {
            "type": "object",
            "required": [
                "permission_id"
            ],
            "properties": {
                "permission_id": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "types.Role": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Admin"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "types.Server": {
            "type": "object",
            "required": [
                "auth_method",
                "default_content_type",
                "http_method",
                "name",
                "url"
            ],
            "properties": {
                "allow_callbacks": {
                    "type": "boolean"
                },
                "allow_copies": {
                    "type": "boolean"
                },
                "auth_method": {
                    "type": "string",
                    "example": "basic"
                },
                "auth_token": {
                    "type": "string"
                },
                "burst": {
                    "type": "integer",
                    "example": 20
                },
                "callback_url": {
                    "type": "string",
                    "example": "https://example.org/callback"
                },
                "cc_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"https://a\"",
                        "\"https://b\"]"
                    ]
                },
                "created": {
                    "type": "string"
                },
                "default_content_type": {
                    "type": "string",
                    "example": "application/json"
                },
                "end_submission_period": {
                    "type": "integer",
                    "example": 2359
                },
                "endpoint_type": {
                    "type": "string"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "{\"Accept\"": "\"application/json\"}"
                    }
                },
                "http_method": {
                    "type": "string",
                    "example": "POST"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "ipaddress": {
                    "type": "string",
                    "example": "10.0.0.12"
                },
                "json_response_xpath": {
                    "type": "string"
                },
                "max_concurrency": {
                    "type": "integer",
                    "example": 2
                },
                "name": {
                    "type": "string",
                    "example": "DHIS2 Main"
                },
                "parse_responses": {
                    "type": "boolean"
                },
                "password": {
                    "type": "string",
                    "example": "district"
                },
                "rate_class": {
                    "type": "string",
                    "example": "server:1"
                },
                "rps": {
                    "type": "number",
                    "example": 10
                },
                "ssl_client_certkey_file": {
                    "type": "string"
                },
                "start_submission_period": {
                    "type": "integer",
                    "example": 0
                },
                "suspended": {
                    "type": "boolean",
                    "example": false
                },
                "system_type": {
                    "type": "string"
                },
                "timeout_ms": {
                    "type": "integer",
                    "example": 10000
                },
                "uid": {
                    "type": "string",
                    "example": "9e2b34f0-1fb2-4a2f-a6d9-0b2a1f92d144"
                },
                "updated": {
                    "type": "string"
                },
                "url": {
                    "type": "string",
                    "example": "https://dhis2.example.org/api/dataValueSets"
                },
                "url_params": {},
                "use_async": {
                    "type": "boolean",
                    "example": true
                },
                "use_ssl": {
                    "type": "boolean",
                    "example": true
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                },
                "xml_response_xpath": {
                    "type": "string"
                }
            }
        },
...
    }
}